+++ examples/example-protocol/src/assets/rust_plugin_test/expected_lib.rs
@@ -1,3 +1,4 @@
+#![allow(unused_imports)]
 #[rustfmt::skip]
 mod export;
 #[rustfmt::skip]