+++ fp-bindgen/src/serializable/serde_json.rs
@@ -1,4 +1,5 @@
 use super::Serializable;
+use crate::prelude::TypeMap;
 use crate::types::{CargoDependency, CustomType, Type, TypeIdent};
 use std::collections::BTreeMap;
 
@@ -16,5 +17,36 @@ impl Serializable for serde_json::Value {
             ts_ty: "any".to_owned(),
             ts_declaration: None,
         })
+    }
+}
+
+impl<K, V> Serializable for serde_json::Map<K, V>
+where
+    K: Serializable,
+    V: Serializable,
+{
+    fn ident() -> TypeIdent {
+        TypeIdent {
+            name: "serde_json::Map".to_string(),
+            generic_args: vec![
+                (TypeIdent::from("K"), vec![]),
+                (TypeIdent::from("V"), vec![]),
+            ],
+            ..Default::default()
+        }
+    }
+
+    fn ty() -> Type {
+        Type::Map(
+            "serde_json::Map".to_owned(),
+            TypeIdent::from("K"),
+            TypeIdent::from("V"),
+        )
+    }
+
+    fn collect_types(types: &mut TypeMap) {
+        types.entry(Self::ident()).or_insert_with(Self::ty);
+        K::collect_types(types);
+        V::collect_types(types);
     }
 }