+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   plugin bindings.
 - Add support for `serde_json::Map` (https://github.com/fiberplane/fp-bindgen/pull/163)
 - Replace the `rust_plugin_module` and `rust_wasmer_runtime_module` annotations with a single `rust_module` annotation.
+- Migrate to Wasmer 3 (https://github.com/fiberplane/fp-bindgen/pull/157)
 
 ## 2.4.0